requirements.txt
# albumentations==1.4.0
# cvzone==1.6.1
# face-detection @ git+https://github.com/elliottzheng/face-detection@786fbab7095623c348e251f1f0a8b323721c6a84
# image==1.5.33
# imutils==0.5.4
# ipython==8.23.0
opencv-python==4.11.0
keras==3.8.0
keras-facenet==0.3.2
l2cs @ git+https://github.com/edavalosanaya/L2CS-Net.git@4a0f978d5b4c426a7d37022d8c927d6ea031dcb6
matplotlib==3.8.4
matplotlib-inline==0.1.7
mediapipe==0.10.21
mtcnn==0.1.1
numpy==1.26.4
pandas==2.2.3
pillow==10.3.0
preprocessing==0.1.13
# requests==2.32.3
# scikit-image==0.24.0
scikit-learn==1.4.2
screeninfo==0.8.1
# tensorflow==2.18.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
utils==1.0.2